% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_inv.R
\name{a_inv}
\alias{a_inv}
\title{Constructs the inverse of the numerator relationship matrix.}
\usage{
a_inv(pedf)
}
\arguments{
\item{pedf}{A potentially reduced data frame for the population but must contain "id", "sire" and "dam".}
}
\value{
The inverse of the numerator relationship matrix.
}
\description{
Does what it says!
}
