% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_mat.R
\name{a_mat}
\alias{a_mat}
\title{Supplies the numerator relationship matrix for a pedigree.}
\usage{
a_mat(pedf)
}
\arguments{
\item{pedf}{A potentially reduced data frame for the population but must contain "id", "sire" and "dam".}
}
\value{
The numerator relationship matrix for the population.
}
\description{
Does what it says!
}
