% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ma_gen.R
\name{add_ma_gen}
\alias{add_ma_gen}
\title{Adds a generation of offspring after selecting mates by maximum avoidance (i.e. minimum coancestry)}
\usage{
add_ma_gen(ddf, hh)
}
\arguments{
\item{ddf}{A data frame with the 9 columns as generated by new_pop}

\item{hh}{Heritability of the trait in the data frame}
}
\value{
An data frame of the same form as the input but extended by a number of offspring deteremined by noff.
   If noff has unequal numbers of offspring from male and female parents, or gives numbers of offspring are not even,
   or the heritability is out of bounds then errors are printed and the input data frame is returned unchanged.
}
\description{
The maximum avoidance is found by a call to amat and using it to evaluate 1000 random swaps. The average
   inbreeding of the offspring and the group coancestry are calculated and printed, together with the deviation from
   random mating. In creating the offspring, the function makes calls to the blup function for calculating the EBV, and
   the a_mat function for calculating the inbreeding coefficient from the leading diagonal.
}
