% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup.R
\name{blup}
\alias{blup}
\title{Calculates BLUP EBVs}
\usage{
blup(ddf, hh)
}
\arguments{
\item{ddf}{The standard data frame of 9 columns containing population data.}

\item{hh}{The heritability of the trait.}
}
\value{
A vector of EBVs
}
\description{
Calculates BLUP EBVs after constructing the inverse of the numerator relationship matrix.
}
