% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_f.R
\name{cost_f}
\alias{cost_f}
\title{Recommends contributions based on relative costs of gain and inbreeding.}
\usage{
cost_f(ddf, tp, tc, cval)
}
\arguments{
\item{ddf}{The data frame of 9 columns holding the population data.}

\item{tp}{The number of parents to be considered for contributions, assumed to be the most recent.}

\item{tc}{The number of offspring to be created.}

\item{cval}{The relative cost of inbreeding which needs to be supplied by the user.}
}
\value{
A logical indicating the success of the algorithm or otherwise.
}
\description{
Carries out the standard penalised contributions algorithm, including constraints on totals contributions
   of males and females, and non-negative contributions. The approximate solution is obtained by the algorithm of
   Meuwissen. It prints out the recommendations after scaling the contributions to give the required total offspring.
   The function checks whether the number of parents to be considered exceeds the population, whether there are both
   males and female parents available. If these checks fail, error messages are printed.
}
