% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_pop.R
\name{new_pop}
\alias{new_pop}
\title{Sets up a base population for selection}
\usage{
new_pop(nn, hh)
}
\arguments{
\item{nn}{Size of base, an even integer}

\item{hh}{Heritability of the simulated trait}
}
\value{
A data frame of 9 labelled columns of integers and doubles. Errors are printed if the size
   of the base is not an even integer, and the heritability is out of bounds.
}
\description{
The first step in the simulation.
}
