% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_sel.R
\name{oc_sel}
\alias{oc_sel}
\title{Recommends contributions for achieving the target group coancestry.}
\usage{
oc_sel(ddf, tp, tc, gcoa)
}
\arguments{
\item{ddf}{The data frame of 9 columns holding the population data.}

\item{tp}{The number of parents to be considered for contributions, assumed to be the most recent.}

\item{tc}{The number of offspring to be created.}

\item{gcoa}{The target group coancestry, which needs to be supplied by the user.}
}
\value{
A logical indicating the success of the algorithm or otherwise.
}
\description{
Carries out the standard optimum contribution algorithm, including constraints on totals contributions
   of males and females, and non-negative contributions. The approximate solution is obtained by the algorithm of
   Meuwissen. It prints out the recommendations after scaling the contributions to give the required total offspring.
   The function checks whether the number of parents to be considered exceeds the population, whether there are both
   males and female parents available, and whether the group coancestry is achievable. If these checks fail, error
   messages are printed.
}
